added profile generation nb together with model
