+++ digital_brain/image-segmentation/.gitignore
@@ -1,0 +1,3 @@
+.DS_Store
+
+assets/*
